requirements.txt
matplotlib>=3.3.4
numpy>=1.22.0
pandas>=1.2.5
PyYAML>=6.0
scipy>=1.8.3
setuptools>=65.5.1
vtk>=9.0.3
pyvista>=0.37.0
ruamel-yaml
regex
pyarrow
meshio 
h5py

# git+git@github.com:wr1/CCBlade.git@v1.3